portfolio-project/backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.12
alembic==1.12.1

# Pydantic
pydantic==2.10.4
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# HTTP Client
httpx==0.24.1

# Redis
redis==5.0.1
hiredis==3.3.0

# Markdown & Code Highlighting
markdown==3.5.1
pygments==2.17.2

# Image Processing
pillow==10.4.0

# Utilities
python-slugify==8.0.1
python-dateutil==2.8.2

# Rate Limiting
slowapi==0.1.9

# Storage (Supabase)
supabase==2.0.3
storage3==0.6.1

# Email
aiosmtplib==3.0.1

# Logging
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
isort==5.13.2
flake8==6.1.0

# Additional dependencies for Windows compatibility
wheel==0.42.0
setuptools==69.0.2
